correction on deleteProfile function

+++ backend/controllers/user.js
@@ -21,10 +21,12 @@ exports.modifyProfile = (req, res, next) => {
     .then(result => result[0])
     .then(user => {
       if (req.file !== undefined) { // L'utilisateur a envoyé une image
-        // Supprimer l'ancienne image de l'API
-        fs.unlink(`./images/${user.picture.split("/images/")[1]}`, (err) => {
-          if (err) throw err;
-        });
+        if (user.picture !== "") {
+          // Supprimer l'ancienne image de l'API
+          fs.unlink(`./images/${user.picture.split("/images/")[1]}`, (err) => {
+            if (err) throw err;
+          });
+        }
         // Mettre la nouvelle
         picture = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
         userMdl.modifyProfile(req.body.firstname, req.body.lastname, req.body.email, picture, req.body.position, req.body.phone, req.params.id)
@@ -47,9 +49,13 @@ exports.deleteProfile = (req, res, next) => {
   userMdl.getProfile(req.params.id)
     .then(result => result[0])
     .then(user => {
-      fs.unlink(`./images/${user.picture.split("/images/")[1]}`, (err) => {
-        if (err) throw err;
-      });
+      // Suppression de l'image
+      // S'il y en a une !
+      if (user.picture !== "") {
+        fs.unlink(`./images/${user.picture.split("/images/")[1]}`, (err) => {
+          if (err) throw err;
+        });
+      }
       userMdl.deleteProfile(req.params.id)
         .then(result => {
           res.status(200)
